+++ src/roq/okx/live.sh
@@ -23,5 +23,4 @@ $PREFIX ./roq-okx \
   --metrics_listen_address 2345 \
   --ws_public_uri "$WS_URI/public" \
   --ws_private_uri "$WS_URI/private" \
-  --ws_books_depth 1 \
   $@