+++ test/json_account.cpp
@@ -130,6 +130,9 @@ TEST_CASE("json_account_parser", "[json_account]") {
     void operator()(Trace<json::MarkPrice const> const &) override { FAIL(); }
     void operator()(Trace<json::Tickers const> const &) override { FAIL(); }
     void operator()(Trace<json::Trades const> const &) override { FAIL(); }
+    void operator()(Trace<json::BboTbt const> const &, [[maybe_unused]] std::string_view const &inst_id) override {
+      FAIL();
+    }
     void operator()(
         Trace<json::BooksL2Tbt const> const &,
         [[maybe_unused]] std::string_view const &inst_id,