+++ test/json_tickers.cpp
@@ -85,6 +85,9 @@ TEST_CASE("json_tickers_parser", "[json_tickers]") {
       CHECK(d0.ts == 1640156513987ms);
     }
     void operator()(Trace<json::Trades const> const &) override { FAIL(); }
+    void operator()(Trace<json::BboTbt const> const &, [[maybe_unused]] std::string_view const &inst_id) override {
+      FAIL();
+    }
     void operator()(
         Trace<json::BooksL2Tbt const> const &,
         [[maybe_unused]] std::string_view const &inst_id,