+++ test/json_trades.cpp
@@ -65,6 +65,9 @@ TEST_CASE("json_trades_parser", "[json_trades]") {
       CHECK(d0.side == json::Side::BUY);
       CHECK(d0.ts == 1640157052811ms);
     }
+    void operator()(Trace<json::BboTbt const> const &, [[maybe_unused]] std::string_view const &inst_id) override {
+      FAIL();
+    }
     void operator()(
         Trace<json::BooksL2Tbt const> const &,
         [[maybe_unused]] std::string_view const &inst_id,